# Use lazy imports for matplotlib in skimage.io