--- a/skimage/io/_plugins/matplotlib_plugin.py
+++ b/skimage/io/_plugins/matplotlib_plugin.py
@@ -1,6 +1,5 @@
 from collections import namedtuple
 import numpy as np
-from mpl_toolkits.axes_grid1 import make_axes_locatable
 import matplotlib.image
 from ...util import dtype as dtypes
 from ...exposure import is_low_contrast
@@ -146,6 +145,7 @@ def imshow(image, ax=None, show_cbar=None, **kwargs):
         The `AxesImage` object returned by `plt.imshow`.
     """
     import matplotlib.pyplot as plt
+    from mpl_toolkits.axes_grid1 import make_axes_locatable
 
     lo, hi, cmap = _get_display_range(image)