## Description

Use lazy imports for matplotlib in skimage.io. This speeds up import for the common case of not using the matplotlib plugin and still gives a speed up when not all the matplotlib functionality is used. For a short script the speedup can be significant.

When using skimage.io the import of matplotlib takes a significant amount time, even in the case that it the matplotlib import is not used.

Using 
```
python -X importtime -m skimage.io 2> import_time_skimage_io.log
tuna import_time_skimage_io.log
```
![Screenshot at 2022-09-28 15-33-11](https://user-images.githubusercontent.com/1113988/192815148-0af5fd5c-f510-4aae-a3ca-94b03050cd24.png)

Shows that 47% of the import time is importing matplotlib (~0.8 out of 1.6 seconds). (I have seen figures down to 37% depending on setup).

With some test scripts, run 10 times in a loop in bash

|Example|Before|After|
|---|---|---|
|import_skimg_io.py | 1.8082s | 0.9405s |
|imread_default.py | 4.9919s | 4.1473s |
|imread_mpl.py | 5.0551s | 4.6022s |

import_skimg_io.py
``` python
import skimage.io
```

imread_default.py
```python
import skimage.io
from skimage._shared.testing import fetch
for _ in range(100):
	img = skimage.io.imread(fetch("data/astronaut.png"))
print(img.shape)
```

imread_mpl.py
```python
import skimage.io
from skimage._shared.testing import fetch
for _ in range(100):
	img = skimage.io.imread(fetch("data/astronaut.png"), plugin="matplotlib")
print(img.shape)
```

Measurements on i5-8250U (restricted to 1GHz to avoid throttling), nvme SSD, on Fedora 36, python 3.10. 

Its not immediately clear how to make a benchmark that measures imports time, but I can look into it if required.

## Checklist

<!-- It's fine to submit PRs which are a work in progress! -->
<!-- But before they are merged, all PRs should provide: -->
- [Docstrings for all functions](https://github.com/numpy/numpy/blob/master/doc/example.py)
- Gallery example in `./doc/examples` (new features only)
- Benchmark in `./benchmarks`, if your changes aren't covered by an
  existing benchmark
- Unit tests
- Clean style in [the spirit of PEP8](https://www.python.org/dev/peps/pep-0008/)
- Descriptive commit messages (see below)

<!-- For detailed information on these and other aspects see -->
<!-- the scikit-image contribution guidelines. -->
<!-- https://scikit-image.org/docs/dev/contribute.html -->

## For reviewers

<!-- Don't remove the checklist below. -->
- Check that the PR title is short, concise, and will make sense 1 year
  later.
- Check that new functions are imported in corresponding `__init__.py`.
- Check that new features, API changes, and deprecations are mentioned in
  `doc/release/release_dev.rst`.
- There is a bot to help automate backporting a PR to an older branch. For
  example, to backport to v0.19.x after merging, add the following in a PR
  comment: `@meeseeksdev backport to v0.19.x`
- To run benchmarks on a PR, add the `run-benchmark` label. To rerun, the label
  can be removed and then added again. The benchmark output can be checked in
  the "Actions" tab.