@samtygier 2022-09-28
The times in the table are from doing
```time (for i in `seq 10`; do python import_skimg_io.py ; done)```
with each of the scripts.

I see there is already a benchmark for imports, so I'll add the module to the param list there ( https://github.com/scikit-image/scikit-image/blob/main/benchmarks/benchmark_import_time.py ).