@samtygier 2022-09-28
I have added `skimage.io` to the import time benchmark. I can see it in the output, but as far as I can understand asv I can't see a difference between my branch and main.

I have also tried converting to the timeraw method, which makes all the import tests faster (I assume because it skips starting a shell), but still I can't see a difference.

The timeraw method is cleaner and probably safer, so I could add a patch so switch the whole test to that. If it is not a problem that it changes existing times.