@lagru 2022-09-29
> The timeraw method is cleaner and probably safer, so I could add a patch so switch the whole test to that. If it is not a problem that it changes existing times.

I'd advocate that change as it's asv's recommended approach. :+1: