@lagru 2022-09-30
Hmm, the benchmark suite still doesn't seem to pick up a difference. Does it show locally for you @samtygier?