@samtygier 2022-10-02
No I've not seen a time difference for this patch with asv. Is it possible that asv already has matplotlib loaded into the interpreter? 

I do see the improvement when I actually run a script that imports `skimage.io`. I've see this on 2 computers and also in docker (to make sure I have clean environments). I can post the docker files if those are useful.