@samtygier 2022-10-02
I tried adding a timeraw test like this:

```python
    def timeraw_import_info(self):
        return """
        import sys
        with open("/tmp/timeraw_import_info", "w") as fh:
            for k,v in sys.modules.items():
                print(k,v, file=fh)
        """
```
In the output I see:
`mpl_toolkits <module 'mpl_toolkits' (<_frozen_importlib_external._NamespaceLoader object at 0x7f80d805feb0>)>`