@hmaarrfk 2022-10-02
When I was doing import time bechmarks, it is import that you do things like commands:

```
python -c "import XXXX"
```
within ASV, as in, start a new interpreter and whatnot.