@hmaarrfk 2022-10-02
to be honest, I am somewhat worried about you trying to help improve this module given that the loose consensus among scikit-image maintainers has been that `skimage.io` provides little improvement over `imageio`. Many of us would like to remove the module `skimage.io` as a whole, that said, it is quite hard.....

For reference if you use:
```
import skimage
import imageio.v3
```
in your code, you could get something more reasonable in terms of import times
![image](https://user-images.githubusercontent.com/90008/193467915-ee333a21-839e-4502-ad65-fd1294e9d32f.png)

https://github.com/imageio/imageio

It is not that I am trying to discourage you from contributing, but rather I do really understand your frustration. Using imageio gives you:

1. Faster import times today. (you don't have to wait for a new release of scikit-image)
2. Avoids churn here.
3. Ensures that you are using what some skimage maintainers think is best practice.