@samtygier 2022-10-02
> When I was doing import time bechmarks, it is import that you do things like commands

The docs suggest that `timeraw` should be good for imports. It also uses `subprocess.check_output` instead of `subprocess.run` so is more robust to errors.

> Many of us would like to remove the module skimage.io as a whole

Should there be a deprecation warning on import?