@hmaarrfk 2022-10-02
> Should there be a deprecation warning on import?

That's more of a discussion for 
https://github.com/scikit-image/scikit-image/issues/5036