@stefanv 2022-10-12
FWIW, I don't think we need benchmarks for changes as these.

Needs a rebase, then should be good to go.