@lagru 2022-10-15
Done. I opted to keep the original way of measuring import time to keep the other benchmarks consistent. Though I tend to agree that the `timeraw_import` way would be the better approach.